Add player lives and gameover score reset
Lives are reduced when the shape collides with the wrong base. Needs to
also include boundary collisions. When the player has no lives left,
the score is reset to 0 and a new shape is chosen.

There is a bug where 2 shapes are chosen after gameover. This will be
looked into for the next commit.